requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
openai==0.28.0
google-api-python-client==2.111.0
google-auth==2.25.2
schedule==1.2.0
python-dotenv==1.0.0
selenium==4.16.0
pytrends==4.9.2
flask==3.0.0
gunicorn==21.2.0